+++ lib/rails_error_dashboard.rb
@@ -17,6 +17,7 @@ require "turbo-rails"
 
 # Core library files
 require "rails_error_dashboard/value_objects/error_context"
+require "rails_error_dashboard/helpers/user_model_detector"
 require "rails_error_dashboard/services/platform_detector"
 require "rails_error_dashboard/services/backtrace_parser"
 require "rails_error_dashboard/services/similarity_calculator"